Renamed yahboom-dogzilla-lite into yahboom-dogzilla-lite-monitor
