+++ software/station/examples/yahboom-dogzilla-lite-monitor/src/yahboom_dogzilla_lite_monitor/display.py
@@ -35,7 +35,7 @@ SPI_WRITE_CHUNK_SIZE = 4096
 GPIO_MAX_NAME_SIZE = 32
 GPIO_HANDLES_MAX = 64
 GPIO_HANDLE_REQUEST_OUTPUT = 1 << 1
-GPIO_CONSUMER_LABEL = b"yahboom-dogzilla-lite"
+GPIO_CONSUMER_LABEL = b"yahboom-dogzilla-lite-monitor"
 
 GPIO_IOCTL_NR_SHIFT = 0
 GPIO_IOCTL_TYPE_SHIFT = 8
