+++ software/station/examples/yahboom-dogzilla-lite-monitor/src/yahboom_dogzilla_lite_monitor/render.py
